# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "User Game API",
    "description": "This is an API for the user games service. The technology used in this API is expressJs and Postgresql database. Made to fulfill the challenge 05 binary academy",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Access to API",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Login Successfully"
                },
                "accessToken": {
                  "type": "string",
                  "example": "[your token]"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid Username Or Password"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error: Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/histories": {
      "get": {
        "tags": [
          "Histories"
        ],
        "summary": "Get all game user history",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "2022-04-08T15:06:30.653Z"
                  },
                  "User": {
                    "type": "object",
                    "properties": {
                      "username": {
                        "type": "string",
                        "example": "botcat"
                      }
                    }
                  },
                  "Activity": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "Login"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthorized : Your Token is Invalid"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "History Not Found!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error: Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get All User",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "raptotor"
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "2022-04-08T15:06:30.316Z"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "2022-04-08T15:06:30.316Z"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthorized : Your Token is Invalid"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User Not Found!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error: Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Add New User",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "mobile_no": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "nickname": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User Has Been Created"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "value": {
                            "type": "string",
                            "example": "any"
                          },
                          "msg": {
                            "type": "string",
                            "example": "Mobile Number is Not valid"
                          },
                          "param": {
                            "type": "string",
                            "example": "mobile_no"
                          },
                          "location": {
                            "type": "string",
                            "example": "body"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthorized : Your Token is Invalid"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error: Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Find User By ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "raptotor"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2022-04-08T15:06:30.316Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2022-04-08T15:06:30.316Z"
                },
                "Biodata": {
                  "type": "object",
                  "properties": {
                    "nickname": {
                      "type": "string",
                      "example": "Rapthor98"
                    },
                    "email": {
                      "type": "string",
                      "example": "siraptorkeren@mail.com"
                    },
                    "mobile_no": {
                      "type": "string",
                      "example": "08888888812"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthorized : Your Token is Invalid"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User Not Found!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error: Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update an exiting User",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "nickname": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "mobile_no": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User Has Been Updated!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthorized : Your Token is Invalid"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User Not Found!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error: Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete a User",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User Has Been Deleted!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthorized : Your Token is Invalid"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User Not Found!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error: Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  }
}